+++ packages/calcite-components/src/components/chip/chip.stories.ts
@@ -8,9 +8,9 @@ import readme from "./readme.md";
 export default {
   title: "Components/Chip",
   parameters: {
-    notes: readme
+    notes: readme,
   },
-  ...storyFilters()
+  ...storyFilters(),
 };
 
 export const simple = (): string => html`