+++ packages/calcite-components/src/components/combobox/resources.ts
@@ -3,5 +3,5 @@ export const ComboboxItemGroup = "CALCITE-COMBOBOX-ITEM-GROUP";
 export const ComboboxChildSelector = `${ComboboxItem}, ${ComboboxItemGroup}`;
 
 export const CSS = {
-  listContainer: "list-container"
+  listContainer: "list-container",
 };