+++ packages/calcite-components/src/components/icon/resources.ts
@@ -1,4 +1,4 @@
 export const CSS = {
   icon: "icon",
-  flipRtl: "flip-rtl"
+  flipRtl: "flip-rtl",
 };