+++ packages/calcite-components/src/components/popover/popover.tsx
@@ -10,7 +10,7 @@ import {
   Prop,
   State,
   VNode,
-  Watch
+  Watch,
 } from "@stencil/core";
 import {
   connectFloatingUI,
@@ -24,7 +24,7 @@ import {
   LogicalPlacement,
   OverlayPositioning,
   ReferenceElement,
-  reposition
+  reposition,
 } from "../../utils/floating-ui";
 import {
   activateFocusTrap,
@@ -32,7 +32,7 @@ import {
   deactivateFocusTrap,
   FocusTrap,
   FocusTrapComponent,
-  updateFocusTrapElements
+  updateFocusTrapElements,
 } from "../../utils/focusTrapComponent";
 import { ARIA_CONTROLS, ARIA_EXPANDED, CSS, defaultPopoverPlacement } from "./resources";
 
@@ -41,7 +41,7 @@ import { guid } from "../../utils/guid";
 import {
   connectOpenCloseComponent,
   disconnectOpenCloseComponent,
-  OpenCloseComponent
+  OpenCloseComponent,
 } from "../../utils/openCloseComponent";
 import { Heading, HeadingLevel } from "../functional/Heading";
 import { Scale } from "../interfaces";
@@ -52,7 +52,7 @@ import {
   disconnectMessages,
   setUpMessages,
   T9nComponent,
-  updateMessages
+  updateMessages,
 } from "../../utils/t9n";
 import { PopoverMessages } from "./assets/popover/t9n";
 import PopoverManager from "./PopoverManager";
@@ -61,7 +61,7 @@ import {
   componentFocusable,
   LoadableComponent,
   setComponentLoaded,
-  setUpLoadableComponent
+  setUpLoadableComponent,
 } from "../../utils/loadable";
 import { createObserver } from "../../utils/observers";
 import { FloatingArrow } from "../functional/FloatingArrow";
@@ -75,7 +75,7 @@ const manager = new PopoverManager();
   tag: "calcite-popover",
   styleUrl: "popover.scss",
   shadow: true,
-  assetsDirs: ["assets"]
+  assetsDirs: ["assets"],
 })
 export class Popover
   implements
@@ -368,7 +368,7 @@ export class Popover
       filteredFlipPlacements,
       offsetDistance,
       offsetSkidding,
-      arrowEl
+      arrowEl,
     } = this;
     return reposition(
       this,
@@ -382,7 +382,7 @@ export class Popover
         offsetDistance,
         offsetSkidding,
         arrowEl,
-        type: "popover"
+        type: "popover",
       },
       delayed
     );
@@ -433,7 +433,7 @@ export class Popover
     const { el, referenceElement, effectiveReferenceElement } = this;
     if (warn && referenceElement && !effectiveReferenceElement) {
       console.warn(`${el.tagName}: reference-element id "${referenceElement}" was not found.`, {
-        el
+        el,
       });
     }
 
@@ -592,7 +592,7 @@ export class Popover
         <div
           class={{
             [FloatingCSS.animation]: true,
-            [FloatingCSS.animationActive]: displayed
+            [FloatingCSS.animationActive]: displayed,
           }}
           // eslint-disable-next-line react/jsx-sort-props
           ref={this.setTransitionEl}
@@ -601,7 +601,7 @@ export class Popover
           <div
             class={{
               [CSS.hasHeader]: !!heading,
-              [CSS.container]: true
+              [CSS.container]: true,
             }}
           >
             {this.renderHeader()}