+++ packages/calcite-components/src/components/text-area/text-area.e2e.ts
@@ -9,7 +9,7 @@ import {
   labelable,
   reflects,
   renders,
-  t9n
+  t9n,
 } from "../../tests/commonTests";
 import { CSS } from "./resources";
 import { html } from "../../../support/formatting";
@@ -23,12 +23,12 @@ describe("calcite-text-area", () => {
     defaults("calcite-text-area", [
       {
         propertyName: "wrap",
-        defaultValue: "soft"
+        defaultValue: "soft",
       },
       {
         propertyName: "scale",
-        defaultValue: "m"
-      }
+        defaultValue: "m",
+      },
     ]);
   });
 
@@ -48,16 +48,16 @@ describe("calcite-text-area", () => {
     reflects("calcite-text-area", [
       {
         propertyName: "columns",
-        value: "10"
+        value: "10",
       },
       {
         propertyName: "rows",
-        value: "50"
+        value: "50",
       },
       {
         propertyName: "scale",
-        value: "s"
-      }
+        value: "s",
+      },
     ]);
   });
 
@@ -76,7 +76,7 @@ describe("calcite-text-area", () => {
     formAssociated("calcite-text-area", {
       testValue: "zion national park",
       expectedSubmitValue: "zion national park",
-      submitsOnEnter: false
+      submitsOnEnter: false,
     });
   });